feat: add separate logic responses
